Move checkAndSave to sync queue
